move helper methods into common file
This allows both specs and features to benefit
from the helper methods.
