+++ features/support/env.rb
@@ -13,6 +13,8 @@ Capybara.javascript_driver = :poltergeist
 Capybara.default_wait_time = 5
 
 class Cucumber::Rails::World
+  require Rails.root.join('spec', 'support', 'common_helper_methods.rb')
+
   def locid_for(resource, *commands, **query_components)
     iri = "#{resource.locid}"
     iri << "///#{commands.join('///')}" if commands.any?
