+++ spec/support/json_schema_matcher.rb
@@ -1,3 +1,5 @@
+require 'rspec/matchers'
+
 RSpec::Matchers.define :match_json_schema do |schema|
   match do |text|
     JSON::Validator.clear_cache
